[JENKINS-49067] Use of @TestExtension instead of @Extension
